Safari-609.1.11.1/LayoutTests/webarchive/test-css-import.html

<html>
<head>
<script>
    if (window.testRunner)
        testRunner.dumpDOMAsWebArchive();
</script>
<link rel="stylesheet" href="resources/test-css-link.css" type="text/css">
<style type="text/css">
    @import url('resources/test-css-import.css');
</style>
</head>
<body>
<div class="blue">This text should be blue.</div>
<div class="red">This text should be red.</div>
<div class="green">This text should be green.</div>
<div class="purple">This text should be purple.</div>
</body>
</html>
